Add files via upload
THIS JOB IS RELATED TO THE FOLLOWING PAPERS: 
- "Modified Connectivity Matrix Algorithm" https://ieeexplore.ieee.org/document/9735116" 
- "Connectivity Matrix Algorithm: A New Optimal Phasor Measurement Unit Placement Algorithm"  https://iopscience.iop.org/article/10.1088/1755-1315/551/1/012008


Connectivity Matrix Algorithm (CMA) is a straightforward algorithm presented to obtain the optimal location of the PMUs in any system. This paper proposed a modification to the connectivity matrix algorithm (MCMA) to consider n-1 criteria. The modified algorithm is valid to get the optimal PMU placement thinking of any level of redundancy in the measurements at any location. The modified algorithm is validated by considering zero injection bus (ZIB) and without for the double redundancy condition. The algorithm applies to different size test systems: IEEE-14 bus, IEEE-30 bus, IEEE 57- bus, and IEEE-118 bus. The computational time of the MCMA is less than 10% of the fastest algorithm in the literature.